Prune the build matrix
Given that this project is being built twice, once for the public tenant
and once for the internal projects, let's make it a bit cheaper and skip
on some configurations which probably do not add that much value,
anyway. I think that three builds are enough for everybody :), and we
still have GCC vs. clang, and regular vs. ASAN vs. ASAN+UBSAN.

This should cut the total wall time of the whole build as measured
across all build nodes by at least 40%.

Change-Id: Ifab38a899bc380f53abc0c1045486948443d8173