Add a build with thread sanitizer and update Netopeer2-server
...and because I'm getting race reports from sysrepo, let's make these
warnings non-fatal so that I can see all of them. Yay!

Also, update Netopeer2-server now that the test suite is stable.

Also, update sysrepo with a timeout simplification patch. This is needed
because `perf_test` is already pretty close to the 300s limit (and it
fails, occasionally).

And update libyang just because I'm already updating the rest of the
NETCONF stack, anyway.

Change-Id: Ib488ba3f1c6157c495ddd47f51ada645e92df636