Migrate from Catch to Doctest
It's something that is quite close to Catch, but it is much, much faster
to compile. Let's give it a try.

There are no matchers and no proper support for data-driven testing, but
I think that we can live without that. Cutting the build time of the
tests by two thirds is worth it, IMHO.

Change-Id: I52f5f9e4a93bd7c6be7d9978b5abcf9d58b0a125