Update the NETCONF stack (and the CI system)
I'm changing the artifact name because the new setup based on Swift
already includes full reposuitory path of the project which produced an
artifact in its name. So let's make this a bit easier to work with.

And finally, the Boost cached location has moved to Swift as well.
Uploading from the CLI can be done like this:

 openstack object create ci-artifacts-public --name 'mirror/buildroot/boost/boost_1_69_0.tar.bz2' boost_1_69_0.tar.bz2

Change-Id: I7cea1f24a6b5a934be3fe4f0cac3b679cc3e33ac