Update the NETCONF stack
There were some test failures related to PID file management in our
containers (they are, unfortunately, sharing /tmp with the host, so it's
persistent). Let's try to redirect sysrepo's PID files and other junk to
a transient location. Ensure that this location is present in the
tarball, but that it also does not keep state between builds.

Here's hoping for some real build VMs, oh yay.

Change-Id: I268da9e16fe8008dd4ccfb7b00e4aa991d7fde9e
Co-authored-by: Jan Kundrát <jan.kundrat@cesnet.cz>