Import pybind11, a Python binding generator
I'm planning to use this one within the autocalibration thingy on top of
netconf-cli. It looks like something more modern than swig (which I had
the pleasure to use on libyang) and boost-python (which prefers Python
2.7 on my machine for some reason, and doesn't recognize the
--with-python3 option that I found on a random stackoverflow post). As a
bonus, it can do std::uniq_ptr, which is nice.

I had to prune their submodules a bit so that it won't try to fetch
extra unrelated content.

Change-Id: I1054f01819ac85f93760622bb47f6bab66d5597e