Building on Fedora 31, too
Fedora 31 is Python3-only, and libredblack's rbgen uses `exec`
statement which won't really work. (There's also a TabError in there,
but that's purely cosmetic.) Libssh also came up a bit more strict
cipher policy settings, hence the libnetconf2 update.

There are new error reports about memory leaks that I was not able to
track down; I don't know if it's due to libev and libssh bugs or whether
the problem is in the callers.

Bug: https://github.com/sysrepo/libredblack/issues/4
Bug: https://github.com/CESNET/libnetconf2/pull/222
Change-Id: I4448c49079f4bd46bba0aef1f3bf558aeb26a64f