CI: reduce number of builds
Instead of building GCC+ASAN+UBSAN and clang+ASAN, let's just go with
clang+ASAN+UBSAN in place of both of these. My reasoning is that we're
already using GCC for the regular build (no ASAN, no UBSAN, no TSAN)
*and* for the ARM build. Let's accompany this two-build GCC with a
two-build clang (ASAN+UBSAN and TSAN). The end result is one build fewer
for leaf projects (and therefore two builds fewer for this multi-tenant
repo).

It looks that the TSAN builds regularly needs 25-30 min to finish. Bump
the allowed time to 45mins so that we have some breathing space.

Adapt to replxx' c55de39c (different options for skipping building of
examples) because that one is still broken for clang+UBSAN.

Change-Id: I11d6f160003fa88139dd834422a8eae661541b70