Use Fedora 32 and its native packages
I was trying to update to the new sysrepo, and that included bumping
buildroot. That, however, resulted in too new `fmt`, which was not
compatible with the version of `spdlog` that we're bundling.

Solve that by using the systemwide spdlog, and while we're at it, do
that for pybind11 as well. I don't think we're ever changing it.

Depends-on: https://gerrit.cesnet.cz/c/ci/zuul-jobs-cesnet/+/3127

Change-Id: Idf19d5f3ed4ed838dcb89a2dc624aa918900ac7b