Use Fedora 32 and its native packages
I was trying to update to the new sysrepo, and that included bumping
buildroot. That, however, resulted in too new `fmt`, which was not
compatible with the version of `spdlog` that we're bundling.

Solve that by using the systemwide spdlog, and while we're at it, do
that for pybind11 as well. I don't think we're ever changing it.

Also try to switch to the systemwide Boost. That's only 1.69 on Fedora
32 (which is older than we have today), but the good thing is that
Fedora 33 has Boost 1.73. Let's see what breaks.

I tried ot do this on legacy sysrepo, but there are some issues with
libev behavior, so I gave up and I'm interested in getting that thing
working with the current stack.

The old LSAN suppressions are not needed anymore, but we have a
wonderful new crash instead, this time in sysrepo's test_modules, and
*only* when running for the first time and with empty repository state.
It segfaults at exit:

 Tracer caught signal 11: addr=0x0 pc=0x4e9218 sp=0x7f9ad95f6d10
 ==25238==LeakSanitizer has encountered a fatal error.
 ==25238==HINT: For debugging, try setting environment variable LSAN_OPTIONS=verbosity=1:log_threads=1
 ==25238==HINT: LeakSanitizer does not work under ptrace (strace, gdb, etc)

...and the backtrace (which is rather hard to obtain because LSAN
doesn't really work under gdb, so the trick is to run it as
`ASAN_OPTIONS=sleep_before_dying=666`, and then attach via `gdb -p
$(pidof test_modules)`, yay!) looks like this:

 (gdb) bt
 #0  __GI___clock_nanosleep (clock_id=clock_id@entry=0, flags=flags@entry=0, req=req@entry=0x7ffc7d3e90a0, rem=rem@entry=0x7ffc7d3e90a0) at ../sysdeps/unix/sysv/linux/clock_nanosleep.c:79
 #1  0x00007fb5d5114157 in __GI___nanosleep (requested_time=requested_time@entry=0x7ffc7d3e90a0, remaining=remaining@entry=0x7ffc7d3e90a0) at nanosleep.c:27
 #2  0x00007fb5d511408e in __sleep (seconds=0) at ../sysdeps/posix/sleep.c:55
 #3  0x00000000004c587b in __asan::AsanDie() ()
 #4  0x00000000004dd56e in __sanitizer::Die() ()
 #5  0x00000000004ec1eb in __lsan::CheckForLeaks() ()
 #6  0x00000000004ec219 in __lsan::DoLeakCheck() ()
 #7  0x00007fb5d50853a7 in __run_exit_handlers (status=0, listp=0x7fb5d5209578 <__exit_funcs>, run_list_atexit=run_list_atexit@entry=true, run_dtors=run_dtors@entry=true) at exit.c:108
 #8  0x00007fb5d5085550 in __GI_exit (status=<optimized out>) at exit.c:139
 #9  0x00007fb5d506d049 in __libc_start_main (main=0x4f3650 <main>, argc=1, argv=0x7ffc7d3e92f8, init=<optimized out>, fini=<optimized out>, rtld_fini=<optimized out>, stack_end=0x7ffc7d3e92e8) at ../csu/libc-start.c:342
 #10 0x000000000041c6ce in _start ()

A TL;DR version is that this is *probably* related to how ASAN guesses
glibc versions, and running it via
`ASAN_OPTIONS=verbosity=2:intercept_tls_get_addr=1` prevents that
segfault.

That was very yummy to debug, of course.

Change-Id: Idf19d5f3ed4ed838dcb89a2dc624aa918900ac7b
Bug: https://github.com/google/sanitizers/issues/1322