Use Fedora 32 and its native packages
I was trying to update to the new sysrepo, and that included bumping
buildroot. That, however, resulted in too new `fmt`, which was not
compatible with the version of `spdlog` that we're bundling.

Solve that by using the systemwide spdlog, and while we're at it, do
that for pybind11 as well. I don't think we're ever changing it.

Also try to switch to the systemwide Boost. That's only 1.69 on Fedora
32 (which is older than we have today), but the good thing is that
Fedora 33 has Boost 1.73. Let's see what breaks.

I tried ot do this on legacy sysrepo, but there are some issues with
libev behavior, so I gave up and I'm interested in getting that thing
working with the current stack.

Change-Id: Idf19d5f3ed4ed838dcb89a2dc624aa918900ac7b