ci: Add a tsan suppresion for glibc tzset_internal
Thread sanitize complains on libc's tzset_internal. It seems that
tzset_internal (used by both libyang and us in
cla-sysrepo/src/utils/time.cpp) is freeing a global variable
(TZ env var) which results in tsan complaining.

Whenever mktime/gmtime/... is run, this code is executed. The code
interprets the TZ env var and IF it was changed (or not yet set),
it stores it inside a global variable. However, this code is executed
*after* __libc_lock_lock (tzset_lock); which is a libc's internal mutex
interface.

We haven't encountered this before because both libyang and our code
from utils were executed from a single thread (Sysrepo callback).

Here is the problematic line [1], called from [2].
It also seems from chromium project encountered the same problem [3].

[1] https://sourceware.org/git/?p=glibc.git;a=blob;f=time/tzset.c;h=16b20b96d3092fae969eebff0c2aaeaf9a9012ca;hb=HEAD#l401
[2] https://sourceware.org/git/?p=glibc.git;a=blob;f=time/tzset.c;h=16b20b96d3092fae969eebff0c2aaeaf9a9012ca;hb=HEAD#l549
[3] https://bugs.chromium.org/p/chromium/issues/detail?id=379738

Change-Id: I1485330c3e950ef97691b81fcaa2d622071563a6