sysrepo: reworked SHM implementation
Most of mutexes are now in stable memory addresses, however, the
subscription one are still not accessed in a safe manner, and TSAN
(correctly) flags that. This requires some updates to the suppressions.
I ended up with something like:

 called_from_lib:libsysrepo.so.5
 race:sr_ptr_
 race:sr_sub
 race:sr_shm
 race:sr_notif_buf

 # our code that's using sysrepo, unfortunately
 race:sr_process_events
 race:sr_apply_changes

...because when I bypass mutex tracking (that I have to, as far as I
know), then there's a plethora of flagged data races. So, a suppression
like that is about as bad as just not using TSAN at all for anything
that passes sysrepo I'm afraid -- note that the `sr_process_events` was
flagged :(.

Also update the rest of the NETCONF stack. These now include some
patches that make it possible to turn on UBSAN enforcing, yay! And while
we're making these changes, let's remove some CMake options which have
not been used for ages in the new sysrepo, and try to see if ASAN breaks
less often on Fedora 33.

TL;DR: fun, fun, fun.

Change-Id: Ic829eb914d32e4352bb09c81c8cb8a9b371eeb2a