Update the NETCONF stack
Nothing super-serious, but hey, why not. Well, actually, a subscription
syncing issue, nice to have that one.

There's been a chain of CI failures in the asan+ubsan job:

 ==4441==Processing thread 4213.
 ==4441==Stack at 0x7ffe1a4d0000-0x7ffe1acd0000 (SP = 0x7ffe1acce2f8).
 ==4441==TLS at 0x7fe5c36d0f40-0x7fe5c36d2240.
 ==4441==DTLS 5 at 0x700006220000030-0xa00006120000032.
 Tracer caught signal 11: addr=0x0 pc=0x4ea6b8 sp=0x7fe5bf2c4d40
 ==4213==LeakSanitizer has encountered a fatal error.
 ==4213==HINT: For debugging, try setting environment variable LSAN_OPTIONS=verbosity=1:log_threads=1
 ==4213==HINT: LeakSanitizer does not work under ptrace (strace, gdb, etc)

This looks pretty much like what's described in commit
db9a4fe2b46f22a8881403f4d1d0677e687f9451, but the funny side is that
this is only happening on one of the two identical build slots -- is
passes for "CzechLight-internal", while failing for the "public" builds.
Anyway, given that the ASAN relaxation was removed in
7570525a6e64a4807e49d65e4515b8725989bac8 without a clear explanation,
let's just put it back here and see if it helps.

Change-Id: I8ca31be79fe1599efc5b816852912194f9eb7cb1