Update dependencies to libyang2 rewrites
This also required upgrading trompeloeil because of:

- https://github.com/rollbear/trompeloeil/issues/264
- a new pretty printer in the C++ bindings

...which means that all consumers now CMake-link via
`trompeloeil::trompeloeil` imported target instead of the old-style
`trompeloeil` thing.

This will however require a fixup to cla-sysrepo due to the changed way
how expectations are managed:

- https://github.com/rollbear/trompeloeil/issues/269

And while we're at this, also sync doctest to the latest release. The
other packages are already either newer-than-the-latest-release or at
the latest-release.

Change-Id: I04b3467bca2748c950e7ab048e0df56970ba0673
Depends-on: https://gerrit.cesnet.cz/c/CzechLight/sysrepo-cpp/+/5522
Co-authored-by: Jan Kundrát <jan.kundrat@cesnet.cz>