CI: try to build leaf projects when updating this repo
Since any backward-incompatible change in upstream projects will very
likely require changes to the consumers, these jobs cannot be voting.

Also try to build the ARM image. This is rather experimental and totally
expected to be unusable at runtime *if* some backwards-incompatible
change among the NETCONF stack, the C++ dependencies, or the "leaf
projects" is introduced.

Change-Id: If264f069c7f939583057570b179b492a8a3284d4
Depends-on: https://gerrit.cesnet.cz/c/CzechLight/rousette/+/5616
Depends-on: https://gerrit.cesnet.cz/c/CzechLight/br2-external/+/5615