Sync the NETCONF stack
This fixes a problem hit by cla-sysrepo's test suite when uninstalling
YANG modules. It was caught by the test suite, but because I've
configured this thing to not hard-fail the build when a depending
project fails (for example, to make it possible to adapt to API
changes), it was merged even though it caused a regression.

This is not the latest & greatest of it all because of some bugs in the
latest releases.

Fixes: 68b26b0 Update the NETCONF stack
Bug: https://github.com/sysrepo/sysrepo/pull/2851
Bug: https://github.com/sysrepo/sysrepo/issues/2852
Change-Id: Ifb7a6bd4e81919e9c76fe0575bc7e18d8b78db1a