Switch to Fedora 38 and update dependencies
This requires temporarily disabling the builds of the depending
projects.

Since the version that's pinned as a submodule wins over a Zuul-level
`override-checkout`, remove that statement and rely on git submodule
pinning.

TSAN got a bit more picky and doesn't play well with SIGEV_THREAD, but I
added an auto-skip of these tests in Netopeer2 upstream (that's for
confirmed-commit and yang-push features).

Change-Id: Ia161f48a4d107abc4d30d0d98e2e5eb270fd929a
Depends-on: https://gerrit.cesnet.cz/c/CzechLight/libyang-cpp/+/5998
Depends-on: https://gerrit.cesnet.cz/c/CzechLight/sysrepo-cpp/+/6204
Depends-on: https://gerrit.cesnet.cz/c/CzechLight/libnetconf2-cpp/+/6206