sync the YANG stack
There was a minor change in netopeer2's installation scripts, there's
now some support for setting the SSH banner. We *could* support that,
but I don't think that it's such a killer feature so I won't bother at
this time. This also means that the czechlight-clearfog build will
(temporarily) fail; a fixup via Buildroot depends on this change in the
br2-external repo.

Change-Id: Ia64d0862d12da93c596c633a0619df818e79502a